backend/requirements.txt
psycopg2-binary==2.9.6
python-dateutil==2.8.2
Babel==2.12.1
Flask==2.2.2  
Flask-BabelEx==0.9.4
flask-cors   
Flask-JWT-Extended==4.4.4
Flask-Login
SQLAlchemy
Flask-SQLAlchemy
Flask-Mail==0.9.1
Flask-Moment==1.0.5
Werkzeug
Flask-WTF==1.1.1
PyJWT==2.5.0
Flask-Bcrypt==1.0.1
bcrypt==3.2.0
Jinja2
WTForms==3.0.1
flask_session
simple-websocket


blinker
click

Flask-SocketIO
itsdangerous
flask-socketio==5.2.0
MarkupSafe
python-engineio
python-socketio
six
Werkzeug